rename infos -> space
