--- symfony/templates/infos.html.twig
@@ -1,58 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block header %}{% endblock %}
-
-{% block body %}
-
-    <br/>
-    <div class="row jumbotron">
-
-        <div class="col-md-12">
-            {% if volunteer.firstName %}
-                <p>{{ 'infos.hello_name'|trans({'%firstName%': volunteer.firstName}) }}</p>
-            {% else %}
-                <p>{{ 'infos.hello_anonymous'|trans }}</p>
-            {% endif %}
-
-            <br/>
-
-            <p>{{ 'infos.context_1'|trans }}</p>
-            <br/>
-            <ul>
-                <li>
-                    <p>{{ 'infos.context_2'|trans }}</p>
-                    <br/>
-                </li>
-                <li>
-                    <p>{{ 'infos.context_3'|trans }}</p>
-                </li>
-            </ul>
-
-            <p>{{ 'infos.context_4'|trans }} {{ 'infos.context_5'|trans }}</p>
-        </div>
-
-        <div class="col-md-6">
-            <br/>
-            <p>{{ 'infos.context_6'|trans }}</p>
-            <img class="img-fluid" src="{{ asset('message.png') }}"/>
-        </div>
-        <div class="col-md-6">
-            <br/>
-            <p>{{ 'infos.context_7'|trans }}</p>
-            <img class="img-fluid" src="{{ asset('reponses.png') }}"/>
-        </div>
-
-        <div class="col-md-12">
-            <br/>
-            <p>{{ 'infos.context_8'|trans }}</p>
-            <br/>
-            <p>{{ 'infos.context_9'|trans }}</p>
-            <p>{{ 'infos.context_10'|trans }}</p>
-            <p>{{ 'infos.context_11'|trans }}</p>
-            <p>{{ 'infos.context_12'|trans }}</p>
-            <p>{{ 'infos.context_13'|trans }}</p>
-        </div>
-
-    </div>
-
-{% endblock %}
