+++ symfony/templates/space/home/index.html.twig
@@ -1,0 +1,58 @@
+{% extends 'base.html.twig' %}
+
+{% block header %}{% endblock %}
+
+{% block body %}
+
+    <br/>
+    <div class="row jumbotron">
+
+        <div class="col-md-12">
+            {% if volunteer.firstName %}
+                <p>{{ 'space.infos.hello_name'|trans({'%firstName%': volunteer.firstName}) }}</p>
+            {% else %}
+                <p>{{ 'space.infos.hello_anonymous'|trans }}</p>
+            {% endif %}
+
+            <br/>
+
+            <p>{{ 'space.infos.context_1'|trans }}</p>
+            <br/>
+            <ul>
+                <li>
+                    <p>{{ 'space.infos.context_2'|trans }}</p>
+                    <br/>
+                </li>
+                <li>
+                    <p>{{ 'space.infos.context_3'|trans }}</p>
+                </li>
+            </ul>
+
+            <p>{{ 'space.infos.context_4'|trans }} {{ 'space.infos.context_5'|trans }}</p>
+        </div>
+
+        <div class="col-md-6">
+            <br/>
+            <p>{{ 'space.infos.context_6'|trans }}</p>
+            <img class="img-fluid" src="{{ asset('message.png') }}"/>
+        </div>
+        <div class="col-md-6">
+            <br/>
+            <p>{{ 'space.infos.context_7'|trans }}</p>
+            <img class="img-fluid" src="{{ asset('reponses.png') }}"/>
+        </div>
+
+        <div class="col-md-12">
+            <br/>
+            <p>{{ 'space.infos.context_8'|trans }}</p>
+            <br/>
+            <p>{{ 'space.infos.context_9'|trans }}</p>
+            <p>{{ 'space.infos.context_10'|trans }}</p>
+            <p>{{ 'space.infos.context_11'|trans }}</p>
+            <p>{{ 'space.infos.context_12'|trans }}</p>
+            <p>{{ 'space.infos.context_13'|trans }}</p>
+        </div>
+
+    </div>
+
+{% endblock %}
