+++ symfony/translations/messages.fr.yml
@@ -419,22 +419,23 @@ maintenance:
       danger: 'Rouge (danger)'
     content: 'Contenu du message (sauts de lignes acceptés)'
 
-infos:
-  hello_name: 'Bievenue %firstName%,'
-  hello_anonymous: 'Bienvenue,'
-  context_1: 'La Croix-Rouge Française possède un peu plus d''un millier de structures ayant des fonctions diverses, notamment dans l''action sociale et le secourisme.'
-  context_2: 'Si par exemple un grave accident a lieu dans un immeuble et des dizaines de personnes se retrouvent à la rue, l''Action Sociale va les prendre en charge, les réchauffer/nourrir si besoin, et trouver des solutions d''urgence.'
-  context_3: 'Ou si par exemple une manifestation déborde et que le SAMU est trop sollicité, ils peuvent faire appel à nous afin qu''on vienne leur donner un coup de main avec nos ambulances.'
-  context_4: 'Pour réunir les effectifs dont une intervention a besoin, les responsables de chaque unité locale doivent efficacement faire appel à leurs bénévoles.'
-  context_5: 'C''est là que RedCall intervient : elle permet à ces responsables de contacter les ressources dont ils ont besoin par SMS ou email.'
-  context_6: 'Ce que le bénévole voit:'
-  context_7: 'Ce que le responsable voit:'
-  context_8: 'Une fois que les équipages sont formés, le responsable peut envoyer un PDF contenant toutes les bonnes informations à sa direction territoriale.'
-  context_9: 'RedCall est actif par défaut pour les DLUS, DTUS et PDT, qui peuvent déclencher leurs structures de rattachement et d''action. S’il y a besoin de droits supplémentaires, la demande peut se faire par email au support Minutis en précisant les éléments suivants :'
-  context_10: '- Nom'
-  context_11: '- Prénom'
-  context_12: '- Nivol'
-  context_13: 'Structures sur lesquels les droits RedCall sont nécessaires.'
+space:
+  home:
+    link: 'Espace bénévole'
+  infos:
+    context_1: 'La Croix-Rouge Française possède un peu plus d''un millier de structures ayant des fonctions diverses, notamment dans l''action sociale et le secourisme.'
+    context_2: 'Si par exemple un grave accident a lieu dans un immeuble et des dizaines de personnes se retrouvent à la rue, l''Action Sociale va les prendre en charge, les réchauffer/nourrir si besoin, et trouver des solutions d''urgence.'
+    context_3: 'Ou si par exemple une manifestation déborde et que le SAMU est trop sollicité, ils peuvent faire appel à nous afin qu''on vienne leur donner un coup de main avec nos ambulances.'
+    context_4: 'Pour réunir les effectifs dont une intervention a besoin, les responsables de chaque unité locale doivent efficacement faire appel à leurs bénévoles.'
+    context_5: 'C''est là que RedCall intervient : elle permet à ces responsables de contacter les ressources dont ils ont besoin par SMS ou email.'
+    context_6: 'Ce que le bénévole voit:'
+    context_7: 'Ce que le responsable voit:'
+    context_8: 'Une fois que les équipages sont formés, le responsable peut envoyer un PDF contenant toutes les bonnes informations à sa direction territoriale.'
+    context_9: 'RedCall est actif par défaut pour les DLUS, DTUS et PDT, qui peuvent déclencher leurs structures de rattachement et d''action. S’il y a besoin de droits supplémentaires, la demande peut se faire par email au support Minutis en précisant les éléments suivants :'
+    context_10: '- Nom'
+    context_11: '- Prénom'
+    context_12: '- Nivol'
+    context_13: 'Structures sur lesquels les droits RedCall sont nécessaires.'
 
 title:
   costs: "Coûts"
